+++ tests/cases/27-p23-context-br.sh
@@ -1,4 +1,4 @@
 # P-23: context.sh không nhận BR-###
 nr p23
 S context.sh BR-001
-xfail P-23 "context.sh BR-001 chạy được (được exit $R)" '[ $R = 0 ]'
+chk "P-23 · context.sh BR-001 chạy được (được exit $R)" '[ $R = 0 ]'